LinuxSDK 2023-10 (LinuxSDK_v2023.10.0.tar.gz)
DTAPI 6.2.0.219
Dta 4.28.10.288
DtaNw 3.5.13.44
DtPcie 3.2.0.354
Dtu 4.15.4.92

+++ LinuxSDK/Drivers/DtPcie/Source/DtCore.c
@@ -474,7 +474,7 @@ DtStatus  DtCore_InitSerialAndHwRev(DtCore*  pCore)
         if (DT_SUCCESS(Status))
         {
             DtDbgOut(MIN, CORE, "Hardware revision set to %d from registry key "
-                                      "'ForcedHardwareRevision'", ForcedHardwareRevision);
+                                 "'ForcedHardwareRevision'", (Int)ForcedHardwareRevision);
             pCore->m_pDevInfo->m_HardwareRevision = (Int)ForcedHardwareRevision;
         }
     }
